+++ kaparoo/filesystem/search/__init__.py
@@ -4,6 +4,7 @@ __all__ = (
     "SPEC_FILE_SUFFIXES",
     "SearchKwargs",
     "Selector",
+    "WalkKwargs",
     "is_spec_file",
     "resolve_selector",
     "search_dirs",
@@ -19,5 +20,5 @@ from kaparoo.filesystem.search.selection import (
     resolve_selector,
     select,
 )
-from kaparoo.filesystem.search.types import SearchKwargs
+from kaparoo.filesystem.search.types import SearchKwargs, WalkKwargs
 from kaparoo.filesystem.search.wrappers import search_dirs, search_files, search_paths